Small change in example

+++ Examples/TcpReassembly/main.cpp
@@ -351,7 +351,7 @@ static void tcpReassemblyMsgReadyCallback(int8_t sideIndex, const TcpStreamData&
 		iter = connMgr->find(tcpData.getConnectionData().flowKey);
 	}
 
-	int side;
+	int8_t side;
 
 	// if the user wants to write each side in a different file - set side as the sideIndex, otherwise write everything to the same file ("side 0")
 	if (GlobalConfig::getInstance().separateSides)
